working on the binary task chabge from "unsigned long int " to "unsigned int"

+++ main.h
@@ -12,7 +12,7 @@ int print_s(char *s);
 
 int print_d_i(long int n);
 
-int print_b(unsigned long int num);
+int print_b(unsigned int num);
 
 /**
   * struct FormatPrinter - code
